Renamed functions from zscan to scan, other miscellaneous.
Changed the names of some methods and variables to reflect the fact that
my experiments are not zscans. Corrected formatting in some places.
Other small changes. FIXME: viewspectra now changes working directory to
parent in order to grab correct files, however this results in issue
with ParticleScanner saving results to parent folder.